refactor(transformers): move transformers to `transformers` dir
BREAKING CHANGE:
One is using all `jest-preset-angular` transformers should change jest config to have:
```
const customTransformers = require('./build/transformers');

module.exports = {
     // [...]
     globals: {
        'ts-jest': {
            tsConfig: '<rootDir>/tsconfig.spec.json',
            stringifyContentPathRegex: '\\.html$',
            astTransformers: {
               before: customTransformers,
            },
         },
     },
}
```

One is using one of `jest-preset-angular` snapshot serializers should change jest config to have:
```
module.exports = {
     // [...]
     globals: {
        'ts-jest': {
            tsConfig: '<rootDir>/tsconfig.spec.json',
            stringifyContentPathRegex: '\\.html$',
            astTransformers: {
               before: [
                  'jest-preset-angular/build/transformers/inline-files'
               ],
            },
         },
     },
}
```
