+++ src/transformers/index.ts
@@ -1,0 +1,1 @@
+export = ['jest-preset-angular/build/transformers/inline-files', 'jest-preset-angular/build/transformers/strip-styles'];
