+++ tsconfig.json
@@ -13,6 +13,6 @@
     "target": "es2015",
     "module": "CommonJS",
     "lib": ["esnext", "dom"],
-    "types": ["node"]
+    "types": ["node", "jest"]
   }
 }
