+++ tsconfig.spec.json
@@ -1,8 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "target": "ES5",
-    "types": ["jest", "node"]
+    "target": "ES5"
   },
   "include": [] // boost test's speed trick
 }
